Fix comment_limit bug.
praw.ini specifies that all a negative value for comment_limit
cause PRAW to use the max for the class. But it only did this
for -1, all other negative values used the reddit default.

Also rearranged the control flow to rise with the value of
comment_limit and use and else. This makes it easier to read
and makes it clear that all values of comment_limit are caught.